can complete a workout with points awarded, no decay yet. also can cancel a workout and revised settings page. new field "timeSinceLastWorkout" added

+++ GymLeague/LeaderboardSection.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct LeaderboardSection {
     let name: String
-    let minPoints: Int
+    let minPoints: Double
 }
 
 let sections = [
